+++ CHANGELOG.md
@@ -9,6 +9,8 @@ See PR links for more information.
 - CHG: changed shrinkler output name from `a.exe.config.shrinkler` to `a.shrinkler_config.exe` so you can use `EXE2ADF` on the resulting shrinkled file
 - CHG: `Makefile`: don't strip symbol names (`-s`) in `elf2hunk` so Shrinkler can display nicer `shrinklerstats`
 - FIX: fixed/optimized `memclr`
+- NEW: ELF disassembly: move cursor with `Shift` to show accumulated cycle counts
+- FIX: flashing document when `Show Source` has been disabled in Profiler `Assembly` view
 
 ## 1.7.6
 - NEW: modified template project to include `-ffunction-sections`, `-fdata-sections`, `--gc-sections` flags and changed assembly files to use unique sections so linker can better strip unused code/data